Stop the name history leaking its transform, and show what MCPvP has
drawNameHistory pushed the panel's transform and then returned early on two
paths -- history still loading, and the player having no previous names --
without popping it. Everything drawn afterwards inherited the panel's origin
and scale, which is why tooltips appeared wildly offset. The common case is
a player with no previous names, so this was hitting most profiles.

MCPvP's rows had no tooltip detail at all: it publishes no rating, no dates
and no per-kit standing -- the same rank and points come back whichever kit
is queried. What it does have is the player's overall placing, so that and
their points are shown instead of "No detail available". Points are carried
as a float now, since MCPvP awards half points and they were being truncated.

+++ src/main/java/com/spog/tiers/data/PlayerTiers.java
@@ -16,7 +16,8 @@ public final class PlayerTiers {
 
 	private String region = "";
 	private int overall = 0;
-	private int points = 0;
+	/** Fractional on MCPvP, where half points are awarded. */
+	private float points = 0.0f;
 
 	public PlayerTiers(TierList source, String name, long fetchedAtMillis) {
 		this.source = source;
@@ -52,11 +53,11 @@ public final class PlayerTiers {
 		this.overall = overall;
 	}
 
-	public int points() {
+	public float points() {
 		return points;
 	}
 
-	public void points(int points) {
+	public void points(float points) {
 		this.points = points;
 	}
